Align PIV/CAC setup-after-sign-in specs to user behavior
changelog: Internal, Automated Testing, Align PIV/CAC setup-after-sign-in specs to user behavior
