+++ spec/support/features/session_helper.rb
@@ -559,7 +559,7 @@ module Features
       click_submit_default
     end
 
-    def stub_piv_cac_service
+    def stub_piv_cac_service(error: nil)
       allow(IdentityConfig.store).to receive(:identity_pki_disabled).and_return(false)
       allow(IdentityConfig.store).to receive(:piv_cac_service_url).
         and_return('http://piv.example.com/')
@@ -570,6 +570,34 @@ module Features
           body: CGI.unescape(request.body.sub(/^token=/, '')),
         }
       end
+
+      stub_request(:post, 'piv.example.com').
+        with(query: hash_including('nonce', 'redirect_uri')).
+        to_return do |request|
+          query = UriService.params(request.uri)
+          {
+            status: 302,
+            headers: {
+              location: UriService.add_params(
+                query['redirect_uri'],
+                token: {
+                  dn: 'C=US, O=U.S. Government, OU=DoD, OU=PKI, CN=DOE.JOHN.1234',
+                  uuid: SecureRandom.uuid,
+                  subject: 'SomeIgnoredSubject',
+                  nonce: query['nonce'],
+                  error:,
+                }.compact.to_json,
+              ),
+            },
+          }
+        end
+    end
+
+    def follow_piv_cac_redirect
+      # RackTest won't do an external redirect to the stubbed PKI service, but we can manually
+      # submit a request to the `current_url` and get the redirect header.
+      redirect_url = Faraday.post(current_url).headers['location']
+      visit redirect_url
     end
 
     def visit_piv_cac_service(idp_url, token_data)
